⚙️Add polygon rotation.

+++ src/Components/TransformComponent.h
@@ -5,9 +5,9 @@
 struct TransformComponent {
     Vec2 position;
     Vec2 scale;
-    double rotation;
+    float rotation;
 
-    TransformComponent(Vec2 position = Vec2::Zero, Vec2 scale = Vec2(1,1), double rotation = 0.0) {
+    TransformComponent(Vec2 position = Vec2::Zero, Vec2 scale = Vec2(1,1), float rotation = 0.0) {
         this->position = position;
         this->scale = scale;
         this->rotation = rotation;
